win: implemented gradients

+++ include/stdlib/_AUX/time_t.h
@@ -1,6 +1,6 @@
 #ifndef _AUX_TIME_T_H
 #define _AUX_TIME_T_H 1
 
-typedef long time_t;
+typedef long long unsigned time_t;
 
 #endif
